Move the preview/Skills.jsx to a separate 'skills' folder.
